VUnit: automatically generate compilation code for system libs
This patch brings functionality of automatic detection of whether to
add or not compilation scripts for the system libraries into the
generate VUnit simulation makefile. The decision is based on presence
of "system" libs in the Manifest module.

If manifest contains "system" key in modules description, this one is
passed to the VUnit make file, which generates compilation code to the
makefile. Currently Altera libs are implemented, but framework permits
easily to extend to other libraries