Fix objdir
OBJDIR\ (ending with backslash does not work, ofcourse)
Spaces in paths does not work, even when quoted, or replaced by %20, '\ ' does not work on windows, so replaces spaces in paths with __

Spaces occurs eg. for some network drives, so if project is located on network drive with spaces, objdir + proj.location => PROJ_OBJ contains spaces,
even if mapped to a drive so visible path does not contain spaces, then git rev-parse --show-toplevel.